+++ src/main/java/com/sparta/quizdemo/config/AsyncConfig.java
@@ -1,4 +1,4 @@
-package com.sparta.quizdemo.sse;
+package com.sparta.quizdemo.config;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.scheduling.annotation.AsyncConfigurer;
