+++ src/main/java/com/sparta/quizdemo/order/service/OrderService.java
@@ -151,8 +151,7 @@ public class OrderService {
                     cartItemRepository.delete(cartItem);
                 }
             }
-            LocalDateTime createdAt = LocalDateTime.now();
-            notificationService.send(user, NotificationType.ORDER_COMPLETED, NotificationType.ORDER_COMPLETED.getMessage(), createdAt);
+            notificationService.send(user, NotificationType.ORDER_COMPLETED, NotificationType.ORDER_COMPLETED.getMessage());
 
             return ResponseEntity.status(HttpStatus.CREATED).body(new ApiResponseDto("결제가 완료 되었습니다.", HttpStatus.CREATED.value()));
         } else {
@@ -222,8 +221,7 @@ public class OrderService {
                 return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(new ApiResponseDto("이미 완료된 주문입니다.", HttpStatus.BAD_REQUEST.value()));
             } else {
                 orderRepository.delete(order);
-                LocalDateTime createdAt = LocalDateTime.now();
-                notificationService.send(user, NotificationType.ORDER_CANCELLED, NotificationType.ORDER_CANCELLED.getMessage(), createdAt);
+                notificationService.send(user, NotificationType.ORDER_CANCELLED, NotificationType.ORDER_CANCELLED.getMessage());
                 return ResponseEntity.status(HttpStatus.OK).body(new ApiResponseDto("주문이 취소 되었습니다.", HttpStatus.OK.value()));
             }
         } else {
