+++ src/main/java/com/sparta/quizdemo/sse/repository/EmitterRepositoryImpl.java
@@ -4,7 +4,6 @@ import lombok.NoArgsConstructor;
 import org.springframework.stereotype.Repository;
 import org.springframework.web.servlet.mvc.method.annotation.SseEmitter;
 
-import java.util.Iterator;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.stream.Collectors;
