+++ src/main/java/com/sparta/quizdemo/sse/service/NotificationListener.java
@@ -1,4 +1,4 @@
-package com.sparta.quizdemo.sse;
+package com.sparta.quizdemo.sse.service;
 
 import com.sparta.quizdemo.sse.dto.NotificationRequestDto;
 import com.sparta.quizdemo.sse.service.NotificationService;
@@ -6,8 +6,6 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.event.TransactionalEventListener;
-
-import java.time.LocalDateTime;
 
 @Component
 @RequiredArgsConstructor
@@ -18,6 +16,6 @@ public class NotificationListener {
     @TransactionalEventListener
     @Async
     public void handleNotification(NotificationRequestDto requestDto) {
-        notificationService.send(requestDto.getUser(), requestDto.getNotificationType(), requestDto.getNotificationType().getMessage(), LocalDateTime.now());
+        notificationService.send(requestDto.getUser(), requestDto.getNotificationType(), requestDto.getNotificationType().getMessage());
     }
 }
