+++ src/main/java/com/sparta/quizdemo/sse/service/NotificationService.java
@@ -3,7 +3,6 @@ package com.sparta.quizdemo.sse.service;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.sparta.quizdemo.common.security.UserDetailsImpl;
-import com.sparta.quizdemo.sse.dto.NotificationRequestDto;
 import com.sparta.quizdemo.sse.dto.NotificationResponseDto;
 import com.sparta.quizdemo.sse.entity.Notification;
 import com.sparta.quizdemo.sse.entity.NotificationType;
@@ -12,7 +11,6 @@ import com.sparta.quizdemo.sse.repository.EmitterRepositoryImpl;
 import com.sparta.quizdemo.sse.repository.NotificationRepository;
 import com.sparta.quizdemo.user.entity.User;
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.MediaType;
 import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -76,7 +74,7 @@ public class NotificationService {
     }
 
     @Async
-    public void send(User user, NotificationType notificationType, String message, LocalDateTime createdAt) {
+    public void send(User user, NotificationType notificationType, String message) {
         Notification notification = notificationRepository.save(createNotification(user, notificationType));
 
         String userId = String.valueOf(user.getId());
