+++ src/main/java/com/sparta/quizdemo/user/service/UserService.java
@@ -124,14 +124,12 @@ public class UserService {
                 () -> new NullPointerException("주소가 존재하지 않습니다.")
         );
 
-        LocalDateTime createdAt = LocalDateTime.now();
-
         //소셜 유저인 경우
         if(updateUser.getSocial() != null){
             updateUser.update(requestDto);
             updateAddress.update(requestDto);
 
-            notificationService.send(user, NotificationType.USER_UPDATE, NotificationType.USER_UPDATE.getMessage(), createdAt);
+            notificationService.send(user, NotificationType.USER_UPDATE, NotificationType.USER_UPDATE.getMessage());
         } else {
             if (!passwordEncoder.matches(requestDto.getPassword(), updateUser.getPassword())) {
                 throw new IllegalArgumentException("비밀번호가 틀립니다.");
@@ -144,13 +142,13 @@ public class UserService {
                 updateUser.update(requestDto, newPassword);
                 updateAddress.update(requestDto);
 
-                notificationService.send(user, NotificationType.USER_UPDATE, NotificationType.USER_UPDATE.getMessage(), createdAt);
+                notificationService.send(user, NotificationType.USER_UPDATE, NotificationType.USER_UPDATE.getMessage());
 
             }else{
                 //일반 정보 수정
                 updateUser.updateUser(requestDto);
                 updateAddress.update(requestDto);
-                notificationService.send(user, NotificationType.USER_UPDATE, NotificationType.USER_UPDATE.getMessage(), createdAt);
+                notificationService.send(user, NotificationType.USER_UPDATE, NotificationType.USER_UPDATE.getMessage());
             }
         }
     }
